Cleanups, AI fixes

+++ src/engine/core/CMap.cpp
@@ -83,7 +83,7 @@ void CMap::refreshStripes()
 // Getters and Setters //
 /////////////////////////
 // returns the tile which is set at the given coordinates
-Uint16 CMap::at(Uint16 x, Uint16 y, Uint16 t)
+Uint16 CMap::at(int x, int y, int t)
 {
 	if(x < m_width && y < m_height )
         return mPlanes[t].getMapDataAt(x,y);
